Split apply.go into file per type of applier
